+++ src/OrchardCore.Modules/Orchard.Alias/Views/_ViewImports.cshtml
@@ -5,6 +5,7 @@
 
 @inherits Orchard.DisplayManagement.Razor.RazorPage<TModel>
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@addTagHelper *, Orchard.Resources
+@addTagHelper *, Orchard.DisplayManagement
+@addTagHelper *, Orchard.ResourceManagement
 
 @using Orchard.Alias.ViewModels
