+++ src/OrchardCore.Modules/Orchard.ContentFields/Views/_ViewImports.cshtml
@@ -6,4 +6,5 @@
 @inherits Orchard.DisplayManagement.Razor.RazorPage<TModel>
 @using Orchard.DisplayManagement
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@addTagHelper *, Orchard.Resources
+@addTagHelper *, Orchard.DisplayManagement
+@addTagHelper *, Orchard.ResourceManagement
