+++ src/OrchardCore.Modules/Orchard.ContentPreview/Views/_ViewImports.cshtml
@@ -6,6 +6,7 @@
 @inherits Orchard.DisplayManagement.Razor.RazorPage<TModel>
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, Orchard.Contents
-@addTagHelper *, Orchard.Resources
+@addTagHelper *, Orchard.DisplayManagement
+@addTagHelper *, Orchard.ResourceManagement
 
 @using Orchard.ContentPreview.ViewModels
