+++ src/OrchardCore.Modules/Orchard.Contents/Views/Content.DetailAdmin.cshtml
@@ -1,4 +1,4 @@
-﻿@using MOrchard.MvcUtilities;
+﻿@using Orchard.Mvc.Utilities;
 @using Orchard.ContentManagement
 @inject IContentManager ContentManager
