+++ src/OrchardCore.Modules/Orchard.Contents/Views/Content.Summary.cshtml
@@ -1,4 +1,4 @@
-﻿@using MOrchard.MvcUtilities;
+﻿@using Orchard.Mvc.Utilities;
 @{
     var contentTypeClassName = ((string)Model.ContentItem.ContentType).HtmlClassify();
 }
