+++ src/OrchardCore.Modules/Orchard.Contents/Views/Content.cshtml
@@ -1,4 +1,4 @@
-﻿@using MOrchard.MvcUtilities;
+﻿@using Orchard.Mvc.Utilities;
 
 @{
