+++ src/OrchardCore.Modules/Orchard.Demo/_ViewImports.cshtml
@@ -5,5 +5,6 @@
 
 @inherits Orchard.DisplayManagement.Razor.RazorPage<TModel>
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@addTagHelper *, Orchard.Resources
+@addTagHelper *, Orchard.DisplayManagement
+@addTagHelper *, Orchard.ResourceManagement
 @addTagHelper *, Orchard.Demo
