+++ src/OrchardCore.Modules/Orchard.Deployment.Remote/Views/_ViewImports.cshtml
@@ -4,4 +4,5 @@
 @using Orchard.Deployment.Remote.ViewModels
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, Orchard.Contents
-@addTagHelper *, Orchard.Resources
+@addTagHelper *, Orchard.DisplayManagement
+@addTagHelper *, Orchard.ResourceManagement
