+++ src/OrchardCore.Modules/Orchard.Deployment/Views/_ViewImports.cshtml
@@ -4,4 +4,5 @@
 @using Orchard.Deployment.ViewModels
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, Orchard.Contents
-@addTagHelper *, Orchard.Resources
+@addTagHelper *, Orchard.DisplayManagement
+@addTagHelper *, Orchard.ResourceManagement
