+++ src/OrchardCore.Modules/Orchard.Flows/Views/_ViewImports.cshtml
@@ -6,7 +6,8 @@
 @inherits Orchard.DisplayManagement.Razor.RazorPage<TModel>
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, Orchard.Contents
-@addTagHelper *, Orchard.Resources
+@addTagHelper *, Orchard.DisplayManagement
+@addTagHelper *, Orchard.ResourceManagement
 
 @using Orchard.ContentManagement;
 @using Orchard.DisplayManagement.Views;
