+++ src/OrchardCore.Modules/Orchard.Layers/Views/_ViewImports.cshtml
@@ -5,7 +5,8 @@
 @inherits Orchard.DisplayManagement.Razor.RazorPage<TModel>
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, Orchard.Contents
-@addTagHelper *, Orchard.Resources
+@addTagHelper *, Orchard.DisplayManagement
+@addTagHelper *, Orchard.ResourceManagement
 
 @using Orchard.ContentManagement;
 @using Orchard.DisplayManagement.Views;
