+++ src/OrchardCore.Modules/Orchard.Modules/Views/_ViewImports.cshtml
@@ -8,4 +8,5 @@
 @using Orchard.Modules.ViewModels;
 
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@addTagHelper *, Orchard.Resources
+@addTagHelper *, Orchard.DisplayManagement
+@addTagHelper *, Orchard.ResourceManagement
