+++ src/OrchardCore.Modules/Orchard.OpenId/Views/_ViewImports.cshtml
@@ -5,6 +5,5 @@
 
 @inherits Orchard.DisplayManagement.Razor.RazorPage<TModel>
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@addTagHelper *, Orchard.ContentManagement.Display
-@addTagHelper *, Orchard.ResourceManagement
 @addTagHelper *, Orchard.DisplayManagement
+@addTagHelper *, Orchard.ResourceManagement
