+++ src/OrchardCore.Modules/Orchard.Resources/Startup.cs
@@ -1,5 +1,9 @@
-﻿using Microsoft.AspNetCore.Modules;
+﻿using System;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Modules;
+using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.DependencyInjection;
+using Orchard.Mvc;
 using Orchard.ResourceManagement;
 
 namespace Orchard.Resources
@@ -9,6 +13,11 @@ namespace Orchard.Resources
         public override void ConfigureServices(IServiceCollection serviceCollection)
         {
             serviceCollection.AddScoped<IResourceManifestProvider, ResourceManifest>();
+        }
+
+        public override void Configure(IApplicationBuilder app, IRouteBuilder routes, IServiceProvider serviceProvider)
+        {
+            serviceProvider.AddTagHelpers("Orchard.ResourceManagement");
         }
     }
 }
