+++ src/OrchardCore.Modules/Orchard.Setup/Views/_ViewImports.cshtml
@@ -1,2 +1,3 @@
 ﻿@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@addTagHelper *, Orchard.Resources
+@addTagHelper *, Orchard.DisplayManagement
+@addTagHelper *, Orchard.ResourceManagement
