+++ src/OrchardCore.Modules/Orchard.Title/_ViewImports.cshtml
@@ -6,4 +6,5 @@
 @inherits Orchard.DisplayManagement.Razor.RazorPage<TModel>
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, Orchard.Contents
-@addTagHelper *, Orchard.Resources
+@addTagHelper *, Orchard.DisplayManagement
+@addTagHelper *, Orchard.ResourceManagement
