+++ src/OrchardCore.Modules/Orchard.Widgets/Views/Widget.cshtml
@@ -1,4 +1,4 @@
-﻿@using MOrchard.MvcUtilities;
+﻿@using Orchard.Mvc.Utilities;
 
 @{
     var contentTypeClassName = ((string)Model.ContentItem.ContentType).HtmlClassify();
