+++ src/OrchardCore.Themes/TheBlogTheme/Views/_ViewImports.cshtml
@@ -6,5 +6,6 @@
 @inherits Orchard.DisplayManagement.Razor.RazorPage<TModel>
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, Orchard.Contents
-@addTagHelper *, Orchard.Resources
+@addTagHelper *, Orchard.DisplayManagement
+@addTagHelper *, Orchard.ResourceManagement
 @addTagHelper *, Orchard.Menu
