+++ src/OrchardCore.Themes/TheTheme/_ViewImports.cshtml
@@ -4,6 +4,6 @@
 
 @inherits Orchard.DisplayManagement.Razor.RazorPage<TModel>
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@addTagHelper *, Orchard.ContentManagement.Display
-@addTagHelper *, Orchard.Resources
+@addTagHelper *, Orchard.DisplayManagement
+@addTagHelper *, Orchard.ResourceManagement
 @addTagHelper *, Orchard.Menu
