+++ src/OrchardCore/Orchard.ResourceManagement.Abstractions/Orchard.ResourceManagement.Abstractions.csproj
@@ -8,6 +8,7 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Mvc.ViewFeatures" Version="1.1.2" />
+    <PackageReference Include="Microsoft.AspNetCore.Razor.Runtime" Version="1.1.1" />
   </ItemGroup>
 
 </Project>
